+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix: Marking vertices/edges as hidden will now update the document in the search index as well as the data store
 * Fix: Fixed an infinite looping problem in the PagingIterable that may have resulted from a malformed ES document
 * Fix: `Elasticsearch5SearchIndex.removeFieldsFromDocument` will not run the removal script in Elasticsearch if there are no properties to remove
+* Fix: Precision and error are now configurable when mapping GeoShape properties. This allows an adjustable trade off between index performance and false positive search results.
 
 # v3.0.0
 * Changed: Removed ES 2 support and replaced it with ES 5 support